Cupcake App with Notification Mockup
Ahem, this is King Explosion, aka Gabriel. It is Saint Patty's Day! Ok to the important thing, this is a mockup of the Cupcake App with the notification on it that tells you that your order is ready. To explain properly, this mockup shows you how the notification works on the cupcake app. It's basically like a text message per se. Well, that's all for the explanation!